+++ config/store.ts
@@ -1,7 +1,17 @@
 import Store from 'electron-store';
 import Conf from 'conf';
+import is from 'electron-is';
 
 import defaults from './defaults';
+
+import { DefaultPresetList, type Preset } from '../plugins/downloader/types';
+
+const getDefaults = () => {
+  if (is.windows()) {
+    defaults.plugins['in-app-menu'].enabled = true;
+  }
+  return defaults;
+};
 
 const setDefaultPluginOptions = (store: Conf<Record<string, unknown>>, plugin: keyof typeof defaults.plugins) => {
   if (!store.get(`plugins.${plugin}`)) {
@@ -10,6 +20,26 @@ const setDefaultPluginOptions = (store: Conf<Record<string, unknown>>, plugin: k
 };
 
 const migrations = {
+  '>=2.1.0'(store: Conf<Record<string, unknown>>) {
+    const originalPreset = store.get('plugins.downloader.preset') as string | undefined;
+    if (originalPreset) {
+      if (originalPreset !== 'opus') {
+        store.set('plugins.downloader.selectedPreset', 'Custom');
+        store.set('plugins.downloader.customPresetSetting', {
+          extension: 'mp3',
+          ffmpegArgs: store.get('plugins.downloader.ffmpegArgs') as string[] ?? DefaultPresetList['mp3 (256kbps)'].ffmpegArgs,
+        } satisfies Preset);
+      } else {
+        store.set('plugins.downloader.selectedPreset', 'Source');
+        store.set('plugins.downloader.customPresetSetting', {
+          extension: null,
+          ffmpegArgs: store.get('plugins.downloader.ffmpegArgs') as string[] ?? [],
+        } satisfies Preset);
+      }
+      store.delete('plugins.downloader.preset');
+      store.delete('plugins.downloader.ffmpegArgs');
+    }
+  },
   '>=1.20.0'(store: Conf<Record<string, unknown>>) {
     setDefaultPluginOptions(store, 'visualizer');
 
@@ -118,7 +148,7 @@ const migrations = {
 };
 
 export default new Store({
-  defaults,
+  defaults: getDefaults(),
   clearInvalidConfig: false,
   migrations,
 });